wnv_hyena_analysis/requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Bioinformatics
biopython>=1.81
biotite>=0.37.0

# Machine learning
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
fastapi>=0.110.0
uvicorn>=0.23.0

# Deep learning genomics
fair-esm>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dimensionality reduction
umap-learn>=0.5.3

# Progress tracking
tqdm>=4.65.0

# Data handling
h5py>=3.8.0
tables>=3.9.0
joblib>=1.3.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Reproducibility
wandb>=0.15.0
